feat(notification): Directly retrieve metadata within the notification provider
Migrate metadata retrieval logic from `GramophonePlaybackService` to `MeiZuLyricsMediaNotificationProvider`.

This simplifies the service implementation and ensures notification construction can access required data more directly.
